% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearSinceBuilt.R
\name{yearSinceBuilt}
\alias{yearSinceBuilt}
\title{yearSinceBuilt}
\usage{
yearSinceBuilt(this_year)
}
\arguments{
\item{this_year}{a integer. This should be the number of the current AD year, for example 2023.}
}
\value{
a selected dataset that only contains year_built and year_since_built in the apt_buildings for visual.
}
\description{
Add new variable year_since_built to apt_buildings dataset. It can also be edited when the new year comes.
For example, just run the function again when the year 2024 arrives and replace the input with 2024.
#'
}
\examples{
# Example usage of the function
library(datateachr)
Example1 <- yearSinceBuilt(2023)
Example2 <- yearSinceBuilt(2024)
}
